feat(isa): Support rv32imac

+++ kernel/defs.h
@@ -31,9 +31,9 @@ struct file*    filealloc(void);
 void            fileclose(struct file*);
 struct file*    filedup(struct file*);
 void            fileinit(void);
-int             fileread(struct file*, uint64, int n);
-int             filestat(struct file*, uint64 addr);
-int             filewrite(struct file*, uint64, int n);
+int             fileread(struct file*, uint32, int n);
+int             filestat(struct file*, uint32 addr);
+int             filewrite(struct file*, uint32, int n);
 
 // fs.c
 void            fsinit(int);
@@ -50,9 +50,9 @@ void            iupdate(struct inode*);
 int             namecmp(const char*, const char*);
 struct inode*   namei(char*);
 struct inode*   nameiparent(char*, char*);
-int             readi(struct inode*, int, uint64, uint, uint);
+int             readi(struct inode*, int, uint32, uint, uint);
 void            stati(struct inode*, struct stat*);
-int             writei(struct inode*, int, uint64, uint, uint);
+int             writei(struct inode*, int, uint32, uint, uint);
 void            itrunc(struct inode*);
 void            ireclaim(int);
 
@@ -70,8 +70,8 @@ void            end_op(void);
 // pipe.c
 int             pipealloc(struct file**, struct file**);
 void            pipeclose(struct pipe*, int);
-int             piperead(struct pipe*, uint64, int);
-int             pipewrite(struct pipe*, uint64, int);
+int             piperead(struct pipe*, uint32, int);
+int             pipewrite(struct pipe*, uint32, int);
 
 // printk.c
 int             printk(char*, ...) __attribute__ ((format (printf, 1, 2)));
@@ -85,7 +85,7 @@ int             kfork(void);
 int             growproc(int);
 void            proc_mapstacks(pagetable_t);
 pagetable_t     proc_pagetable(struct proc *);
-void            proc_freepagetable(pagetable_t, uint64);
+void            proc_freepagetable(pagetable_t, uint32);
 int             kkill(int);
 int             killed(struct proc*);
 void            setkilled(struct proc*);
@@ -96,11 +96,11 @@ void            scheduler(void) __attribute__((noreturn));
 void            sched(void);
 void            sleep(void*, struct spinlock*);
 void            userinit(void);
-int             kwait(uint64);
+int             kwait(uint32);
 void            wakeup(void*);
 void            yield(void);
-int             either_copyout(int user_dst, uint64 dst, void *src, uint64 len);
-int             either_copyin(void *dst, int user_src, uint64 src, uint64 len);
+int             either_copyout(int user_dst, uint32 dst, void *src, uint32 len);
+int             either_copyin(void *dst, int user_src, uint32 src, uint32 len);
 void            procdump(void);
 
 // swtch.S
@@ -132,9 +132,9 @@ char*           strncpy(char*, const char*, int);
 // syscall.c
 void            argint(int, int*);
 int             argstr(int, char*, int);
-void            argaddr(int, uint64 *);
-int             fetchstr(uint64, char*, int);
-int             fetchaddr(uint64, uint64*);
+void            argaddr(int, uint32 *);
+int             fetchstr(uint32, char*, int);
+int             fetchaddr(uint32, uint32*);
 void            syscall();
 
 // trap.c
@@ -153,22 +153,22 @@ void            uartputc_sync(int);
 // vm.c
 void            kvminit(void);
 void            kvminithart(void);
-void            kvmmap(pagetable_t, uint64, uint64, uint64, int);
-int             mappages(pagetable_t, uint64, uint64, uint64, int);
+void            kvmmap(pagetable_t, uint32, uint32, uint32, int);
+int             mappages(pagetable_t, uint32, uint32, uint32, int);
 pagetable_t     uvmcreate(void);
-uint64          uvmalloc(pagetable_t, uint64, uint64, int);
-uint64          uvmdealloc(pagetable_t, uint64, uint64);
-int             uvmcopy(pagetable_t, pagetable_t, uint64);
-void            uvmfree(pagetable_t, uint64);
-void            uvmunmap(pagetable_t, uint64, uint64, int);
-void            uvmclear(pagetable_t, uint64);
-pte_t *         walk(pagetable_t, uint64, int);
-uint64          walkaddr(pagetable_t, uint64);
-int             copyout(pagetable_t, uint64, char *, uint64);
-int             copyin(pagetable_t, char *, uint64, uint64);
-int             copyinstr(pagetable_t, char *, uint64, uint64);
-int             ismapped(pagetable_t, uint64);
-uint64          vmfault(pagetable_t, uint64, int);
+uint32          uvmalloc(pagetable_t, uint32, uint32, int);
+uint32          uvmdealloc(pagetable_t, uint32, uint32);
+int             uvmcopy(pagetable_t, pagetable_t, uint32);
+void            uvmfree(pagetable_t, uint32);
+void            uvmunmap(pagetable_t, uint32, uint32, int);
+void            uvmclear(pagetable_t, uint32);
+pte_t *         walk(pagetable_t, uint32, int);
+uint32          walkaddr(pagetable_t, uint32);
+int             copyout(pagetable_t, uint32, char *, uint32);
+int             copyin(pagetable_t, char *, uint32, uint32);
+int             copyinstr(pagetable_t, char *, uint32, uint32);
+int             ismapped(pagetable_t, uint32);
+uint32          vmfault(pagetable_t, uint32, int);
 
 // plic.c
 void            plicinit(void);
