+++ kernel/exec.c
@@ -7,7 +7,7 @@
 #include "defs.h"
 #include "elf.h"
 
-static int loadseg(pde_t *, uint64, struct inode *, uint, uint);
+static int loadseg(pde_t *, uint32, struct inode *, uint, uint);
 
 // map ELF permissions to PTE permission bits.
 int
@@ -29,7 +29,7 @@ kexec(char *path, char **argv)
 {
   char *s, *last;
   int i, off;
-  uint64 argc, sz = 0, sp, ustack[MAXARG], stackbase;
+  uint32 argc, sz = 0, sp, ustack[MAXARG], stackbase;
   struct elfhdr elf;
   struct inode *ip;
   struct proghdr ph;
@@ -46,7 +46,7 @@ kexec(char *path, char **argv)
   ilock(ip);
 
   // Read the ELF header.
-  if (readi(ip, 0, (uint64)&elf, 0, sizeof(elf)) != sizeof(elf))
+  if (readi(ip, 0, (uint32)&elf, 0, sizeof(elf)) != sizeof(elf))
     goto bad;
 
   // Is this really an ELF file?
@@ -58,7 +58,7 @@ kexec(char *path, char **argv)
 
   // Load program into memory.
   for (i = 0, off = elf.phoff; i < elf.phnum; i++, off += sizeof(ph)) {
-    if (readi(ip, 0, (uint64)&ph, off, sizeof(ph)) != sizeof(ph))
+    if (readi(ip, 0, (uint32)&ph, off, sizeof(ph)) != sizeof(ph))
       goto bad;
     if (ph.type != ELF_PROG_LOAD)
       continue;
@@ -68,7 +68,7 @@ kexec(char *path, char **argv)
       goto bad;
     if (ph.vaddr % PGSIZE != 0)
       goto bad;
-    uint64 sz1;
+    uint32 sz1;
     if ((sz1 = uvmalloc(pagetable, sz, ph.vaddr + ph.memsz,
                         flags2perm(ph.flags))) == 0)
       goto bad;
@@ -81,13 +81,13 @@ kexec(char *path, char **argv)
   ip = 0;
 
   p = myproc();
-  uint64 oldsz = p->sz;
+  uint32 oldsz = p->sz;
 
   // Allocate some pages at the next page boundary.
   // Make the first inaccessible as a stack guard.
   // Use the rest as the user stack.
   sz = PGROUNDUP(sz);
-  uint64 sz1;
+  uint32 sz1;
   if ((sz1 = uvmalloc(pagetable, sz, sz + (USERSTACK + 1) * PGSIZE, PTE_W)) ==
       0)
     goto bad;
@@ -112,11 +112,11 @@ kexec(char *path, char **argv)
   ustack[argc] = 0;
 
   // push a copy of ustack[], the array of argv[] pointers.
-  sp -= (argc + 1) * sizeof(uint64);
+  sp -= (argc + 1) * sizeof(uint32);
   sp -= sp % 16;
   if (sp < stackbase)
     goto bad;
-  if (copyout(pagetable, sp, (char *)ustack, (argc + 1) * sizeof(uint64)) < 0)
+  if (copyout(pagetable, sp, (char *)ustack, (argc + 1) * sizeof(uint32)) < 0)
     goto bad;
 
   // a0 and a1 contain arguments to user main(argc, argv)
@@ -155,11 +155,11 @@ bad:
 // and the pages from va to va+sz must already be mapped.
 // Returns 0 on success, -1 on failure.
 static int
-loadseg(pagetable_t pagetable, uint64 va, struct inode *ip, uint offset,
+loadseg(pagetable_t pagetable, uint32 va, struct inode *ip, uint offset,
         uint sz)
 {
   uint i, n;
-  uint64 pa;
+  uint32 pa;
 
   for (i = 0; i < sz; i += PGSIZE) {
     pa = walkaddr(pagetable, va + i);
@@ -169,7 +169,7 @@ loadseg(pagetable_t pagetable, uint64 va, struct inode *ip, uint offset,
       n = sz - i;
     else
       n = PGSIZE;
-    if (readi(ip, 0, (uint64)pa, offset + i, n) != n)
+    if (readi(ip, 0, (uint32)pa, offset + i, n) != n)
       return -1;
   }
